pyproject.toml
[tool.poetry]
name = "multi_ternant"
version = "0.1.0"
description = ""
authors = ["Eyong Kevin Enowanyo <tonyparkerkenz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.7"
djangorestframework = "^3.12.4"
mixer = "^7.1.2"
pudb = "^2021.1"
psycopg2 = "^2.9.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
